requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1
confluent-kafka==2.3.0
psycopg2==2.9.9
daphne==4.0.0
